# swagger.yml
openapi: 3.0.0
info:
  title: "RP Flask REST API"
  description: "An API about todo Application"
  version: "1.0.0"
servers:
  - url: "/api"
components:
  schemas:
...
    EditProfile:
      type: object
      properties:
        username:
          type: string
        about_me:
          type: string
      required:
        - username
        - about_me
...
    ConnectionData:
      type: object
      properties:
        username:
          type: string
        to_add:
          type: string
        c_type:
          type: string
          enum:
            - clients
            - colleagues
            - mentors
      required:
        - username
        - to_add
        - c_type
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: jwt_manager.security_handler
paths:
...
  /user/profile/update:
    put:
      operationId: profile.update_profile
      tags:
        - Profile
      summary: "Updates a user profile"
      requestBody:
          description: "Updates a user profile"
          required: True
          content:
            application/x-www-form-urlencoded:
              schema:
                x-body-name: "updateprofile"
                $ref: "#/components/schemas/EditProfile"
      security:
        - jwt_auth: ['secret']
      responses:
        '200':
          description: Profile Updated successfully
  /user/{username}/profile/get:
    get:
      operationId: profile.get_profile
      tags:
        - Profile
      summary: "Retrieves a user profile"
      parameters:
        - name: username
          in: path
          description: The username of the profile to retrieve
          required: true
          schema:
            type: string
      security:
        - jwt_auth: []
      responses:
        '200':
          description: Profile Retrieved successfully
  /user/{username}/profile/delete:
    delete:
      operationId: profile.delete_profile
      tags:
        - Profile
      summary: "Deletes a user profile"
      parameters:
        - name: username
          in: path
          description: The username of the profile to delete
          required: true
          schema:
            type: string
      security:
        - jwt_auth: []
      responses:
        '200':
          description: Profile has been deleted successfully
...
  /user/connections/add:
    post:
      operationId: connections.add_connection
      tags:
        - Users
      summary: "Adds a connection to a user"
      requestBody:
        description: "Add connection"
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              x-body-name: "connect"
              $ref: "#/components/schemas/ConnectionData"
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: connection added successfully
  /user/{username}/connections:
    get:
      operationId: connections.get_connected_users
      tags:
        - Users
      summary: "Gets all connected users"
      parameters:
        - name: username
          in: path
          description: The username of the account
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: connections retrieved successfully
  /user/{username}/connections/{to_remove}:
    put:
      operationId: connections.remove_connected_user
      tags:
        - Users
      summary: "Removes a connected user"
      parameters:
        - name: username
          in: path
          description: The username of the account
          required: true
          schema:
            type: string
        - name: to_remove
          in: path
          description: The username of connection to remove
          required: true
          schema:
            type: string
      security:
        - jwt_auth: [ ]
      responses:
        '200':
          description: connection removed successfully
...